+++ previous-versions/commerce/pkg/models/operations/getcustomer.go
@@ -8,7 +8,9 @@ import (
 )
 
 type GetCustomerRequest struct {
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
 	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 	// Unique identifier for a customer.
 	CustomerID string `pathParam:"style=simple,explode=false,name=customerId"`
@@ -36,13 +38,16 @@ func (o *GetCustomerRequest) GetCustomerID() string {
 }
 
 type GetCustomerResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// OK
 	Customer *shared.Customer
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetCustomerResponse) GetContentType() string {