+++ previous-versions/commerce/pkg/models/operations/getproductcategory.go
@@ -8,7 +8,9 @@ import (
 )
 
 type GetProductCategoryRequest struct {
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
 	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 	// Unique identifier for a product.
 	ProductID string `pathParam:"style=simple,explode=false,name=productId"`
@@ -36,13 +38,16 @@ func (o *GetProductCategoryRequest) GetProductID() string {
 }
 
 type GetProductCategoryResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
 	// OK
 	ProductCategory *shared.ProductCategory
-	StatusCode      int
-	RawResponse     *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetProductCategoryResponse) GetContentType() string {