+++ previous-versions/commerce/RELEASES.md
@@ -331,3 +331,13 @@ Based on:
 - [go v0.24.0] previous-versions/commerce
 ### Releases
 - [Go v0.24.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/commerce/v0.24.0 - previous-versions/commerce
+
+## 2023-10-04 00:17:43
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.93.2 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.25.0] previous-versions/commerce
+### Releases
+- [Go v0.25.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/commerce/v0.25.0 - previous-versions/commerce